fix: tf2 lite nano camera support

+++ donkeycar/parts/tflite.py
@@ -1,4 +1,5 @@
 import os
+import numpy as np
 import tensorflow as tf
 
 from donkeycar.parts.keras import KerasPilot
@@ -55,7 +56,7 @@ class TFLitePilot(KerasPilot):
         self.input_shape = self.input_details[0]['shape']
     
     def inference(self, img_arr, other_arr):
-        input_data = img_arr.reshape(self.input_shape)
+        input_data = np.float32(img_arr.reshape(self.input_shape))
         self.interpreter.set_tensor(self.input_details[0]['index'], input_data)
         self.interpreter.invoke()
